Add type hints to mtga_follower
[Python Type Hints](https://docs.python.org/3/library/typing.html) add
clarity to the ambiguity of variable, argument, and function return
types in python. They can be added without having any effect on the code
runing at all.

There are tools, such as [mypy](https://mypy-lang.org/) that can be added
to the linting and CI process to validate that the code is utilizing
types correct and helps prevent subtle bugs. This may be added in a
subsequent change-set.